backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy>=2.0
aiosqlite
passlib[bcrypt]
python-jose[cryptography]

